[#4] Content-assist for Swagger
Refactoring of the code related to code assist to make it more readable.
Main changes is that completions are now obtain solely by a single call made to SwaggerProposalProvider.getCompletionProposals(), no calls to
SwaggerSchema are necessary.
The codebase of SwaggerProposalProvider has also been reduced significantly, to avoid duplicate code. The main change is that the method createProposals will
return a list of JsonNode being the proposals. All proposals have the same format, meaning that a single method can transform that list into a list of ICompletionProposal.