[#30] Issues with Content-assist
This commit includes changes to support addition in content assist of references inside the same document and for resolution
of external schemas during generation of content proposals.

It also includes a refactoring of classes that manipulates JSON. These have been moved to a package named `json`.
It introduces those new classes:
- JsonSchemaManager that resolves and keep in memory the used schema
- SchemaDefinition holds data necessary to create a content assist proposal from a schema definition
- SchemaDefinitionProvider contains methods previously in SwaggerSchema usefull for traversing a given schema and find schema definitions that
matches a given path.