#104 - delete paths template
I think the paths template doesn't really add value.  We don't want people to use this very minimal resource template that's included; we want them to use our object resource and collection resource templates.  When you remove that, all that's left is the `paths:` key, which is already presented as a keyword suggestion.