[#117] Warn on duplicate properties
Rewrite the implementation of the duplicate keys validation method to avoid excessive iteration. The method now uses a multi map of pair<Node, String> x Set<Node>. With the pair left element being a parent node and the pair right element being a key value, the Set of node contains all children of the parent node being a key node and whose value matches the pair right element. Resolution of duplicates is done only once after all of them are collected in the map.