[#40] Implement Quick Outline
QuickOutline and Outline now uses a unified model to get access to the content of the document. The model provides a representation of the
document in form of nodes (object - array - value). Each node can be retrieve via a pointer (JSON pointer) or it's position inside the document.
Each node stores it's start and end position. Each node has access to it's children and parent, making it useful to be used in a tree content provider,
and can provide a label useful for label providers.

This commit also introduces better labels for the (quick)outline. Those labels are now derived from the type of elements.