[#239] Swagger Spec pasted into editor shows spurious warnings until saved
Resolution of JSON references will be done, if references are locals, by inspecting the content of the current documents and not by loading the content of the file on disk. This will ensure that validation of references will be correct when new content is copy/paste inside the current editor.