175 - Handle edge case in setting selection
If an object is selected and that object is fully contained in the final
line of the file and does not have a line terminator, the code for
computing the range to be highlighed can't ask for the position of the
next line without throwing BadLocationException. That case is now
checked for explicitly.