#257 - prevent potential
The modified code had the assumption that this would always be the last
oneOf element. This assumption can be unwittingly violated in some
future code, leading to sporadic failures. This change protects us from
this kind of problems in future. The current code is slightly slower,
but this only happens when preferences change.