ZEN-3411 - SwaggerEditor: changes are not saved on closing file
Problem:
Eclipse ensures that if a view needs to be saved on closing, it’s saved
*before* it’s closed:
org.eclipse.e4.ui.workbench.renderers.swt.StackRenderer.closePart(Widget,
boolean) 
{code}

Why is the order important? Because #hidePart() disposes the
documentProvider which is needed for proper save:
org.eclipse.ui.texteditor.AbstractTextEditor.doSave(IProgressMonitor)
So if we save first and then close the editor, it’s OK. But not the vice
versa.

In our case, we first close the editor and then try to save it. And it
does not work. While doSave() is still called *before* #dispose(), it
wraps its contents inside a SafeWorkspaceJob which schedules it to be
executed later - after the close in our case. 
Why do we need to wrap it in a Job? A job is needed to batch all marker
changes into a single delta for ZEN-2736 Refresh live views on swagedit
error changes