[WIP] - Performance improvements by reducing the number of steps required to parse a document
These changes aim to improve the overall performance of document parsing by reducing the number of steps involved when a document is validated or a call to content assist is made. This is done by using a single parser (whereas 3 parses where done previously using the SnakeYAML parser then the Jakcson/YAML parser and then a custom Model parser). This single parser is an extension of the Jackson/YAML parser that keeps track of JSONNode locations. Those locations are used to locate the position of elements during content assist or when a validation is executed.