[ZEN-3017] OAS3 Level 3 Validation: Schema Objects, examples, alignment with OAS conformance suite
This adds normalization of schemas present in Swagger/OpenAPI. When an example is detected, the object being it's schema is passed through the normalizer to be transformed into a single object where all references are resolved.
This also extends the validation preferences for Swagger/OpenAPI to add an option to enable example validation. This option is set to false by default.