[ZEN-4990] Remove requirement for a KZOP update site
We strip the versions of the dependencies that are copied into the lib folder so that build scripts don't have to modify the MANIFEST file.
This file should still be updated and committed if any change requires it. For example if a dependency is added or removed.

The maven build will also add a file named classPath.txt into the lib folder. This is done by the maven dependency plugin. This file will contain dependency versions that are needed by the gentemplate wizard to set versions of gentemplate projects dependencies (such as genflow, kzop, rapidml).